rename change_detection.rs to query_filter.rs
