+++ src/lib.rs
@@ -262,8 +262,8 @@ use bevy::{
 mod tests;
 
 pub mod all;
-mod change_detection;
 pub mod one;
+mod query_filter;
 
 pub use all::*;
 pub use one::*;
